Change Entity Namespace
