+++ system/Entity/Cast/DatetimeCast.php
@@ -9,18 +9,17 @@
  * file that was distributed with this source code.
  */
 
-namespace CodeIgniter\EntityCast;
+namespace CodeIgniter\Entity\Cast;
 
-use Exception;
-use DateTime;
 use CodeIgniter\I18n\Time;
+use DateTime;
+use Exception;
 
 /**
- * Class CastAsDatetime
+ * Class DatetimeCast
  */
-class CastAsDatetime extends AbstractCast
+class DatetimeCast extends BaseCast
 {
-
 	/**
 	 * @inheritDoc
 	 *
