+++ system/Entity/Cast/TimestampCast.php
@@ -9,16 +9,15 @@
  * file that was distributed with this source code.
  */
 
-namespace CodeIgniter\EntityCast;
+namespace CodeIgniter\Entity\Cast;
 
-use CodeIgniter\Exceptions\CastException;
+use CodeIgniter\Entity\Exceptions\CastException;
 
 /**
- * Class CastAsTimestamp
+ * Class TimestampCast
  */
-class CastAsTimestamp extends AbstractCast
+class TimestampCast extends BaseCast
 {
-
 	/**
 	 * @inheritDoc
 	 */
