+++ tests/_support/Entity/Cast/NotExtendsBaseCast.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Tests\Support\Entity\Cast;
+
+class NotExtendsBaseCast
+{
+
+}
