--- tests/_support/EntityCast/NotExtendsAbstractCast.php
@@ -1,8 +1,0 @@
-<?php
-
-
-namespace Tests\Support\EntityCast;
-
-class NotExtendsAbstractCast
-{
-}
