+++ tests/_support/Models/SimpleEntity.php
@@ -1,6 +1,8 @@
-<?php namespace Tests\Support\Models;
+<?php
 
-use CodeIgniter\Entity;
+namespace Tests\Support\Models;
+
+use CodeIgniter\Entity\Entity;
 
 /**
  * Class SimpleEntity
@@ -12,5 +14,4 @@ use CodeIgniter\Entity;
  */
 class SimpleEntity extends Entity
 {
-
 }
